+++ sdk/compute/arm-compute/src/models/parameters.ts
@@ -110,7 +110,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-07-01",
+    defaultValue: "2023-09-01",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -189,6 +189,26 @@ export const vmScaleSetName: OperationURLParameter = {
   mapper: {
     serializedName: "vmScaleSetName",
     required: true,
+    type: {
+      name: "String"
+    }
+  }
+};
+
+export const ifMatch: OperationParameter = {
+  parameterPath: ["options", "ifMatch"],
+  mapper: {
+    serializedName: "If-Match",
+    type: {
+      name: "String"
+    }
+  }
+};
+
+export const ifNoneMatch: OperationParameter = {
+  parameterPath: ["options", "ifNoneMatch"],
+  mapper: {
+    serializedName: "If-None-Match",
     type: {
       name: "String"
     }
@@ -966,7 +986,7 @@ export const diskName: OperationURLParameter = {
 export const apiVersion1: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-04-02",
+    defaultValue: "2022-07-02",
     isConstant: true,
     serializedName: "api-version",
     type: {