+++ sdk/compute/arm-compute/src/operations/virtualMachineScaleSetVMs.ts
@@ -1373,7 +1373,12 @@ const updateOperationSpec: coreClient.OperationSpec = {
     Parameters.vmScaleSetName,
     Parameters.instanceId
   ],
-  headerParameters: [Parameters.accept, Parameters.contentType],
+  headerParameters: [
+    Parameters.accept,
+    Parameters.contentType,
+    Parameters.ifMatch,
+    Parameters.ifNoneMatch
+  ],
   mediaType: "json",
   serializer
 };